+++ opentelemetry-user-events-logs/src/logs/mod.rs
@@ -11,6 +11,7 @@ pub trait UserEventsLoggerProviderBuilderExt {
     /// Adds a user event exporter to the logger provider builder with the given provider name.
     ///
     /// The provider name must:
+    /// - Not be empty.
     /// - Be less than 234 characters.
     /// - Contain only ASCII letters, digits, and the underscore (`'_'`) character.
     /// - Be short, human-readable, and unique enough to avoid conflicts with other provider names.
@@ -39,8 +40,8 @@ impl UserEventsLoggerProviderBuilderExt for LoggerProviderBuilder {
     fn with_user_events_exporter(self, provider_name: &str) -> Self {
         match UserEventsExporter::new(provider_name) {
             Ok(exporter) => {
-                let reenterant_processor = ReentrantLogProcessor::new(exporter);
-                self.with_log_processor(reenterant_processor)
+                let reentrant_processor = ReentrantLogProcessor::new(exporter);
+                self.with_log_processor(reentrant_processor)
             }
             Err(e) => {
                 otel_warn!(name: "User_Events.Exporter.CreationFailed", reason = &e);